♻️ bot/_legacy 에서 startup + channels 모듈 분리 (P0-Q step 6)
변경 이유:
- _startup_messages 와 채널 이름 정규화 헬퍼는 bot/_legacy.py 안에서 가장 위쪽 부트 시점에만 쓰여 책임 경계가 분명함
- _normalize_channel_name / _channel_target_text 은 향후 추가 분해 모듈 (engineering_handlers / message_routing) 도 함께 import 할 공용 유틸이라 leaf 위치(channels.py) 로 격리

주요 변경 사항:
- src/yule_orchestrator/discord/bot/channels.py 신설 (22줄)
  · _normalize_channel_name, _channel_target_text — 순수 문자열 유틸
- src/yule_orchestrator/discord/bot/startup.py 신설 (140줄)
  · _startup_messages, _channel_configuration_warnings, _channel_overlap_warnings — DiscordBotConfig + now → 부트 로그 라인 리스트
- _legacy.py 에서 동일 정의 5개 삭제 + 각 모듈에서 re-import (facade attribute 유지)

비고:
- 4672 PASS / 1 skipped — 회귀 0건
- 외부 import 경로 (bot 패키지 facade) 그대로 유지
- 다음 분해 대상: engineering_conversation_bridge / forum_factories / message_routing
